pyproject.toml
[project]
name = "libreoffice-mcp"
version = "0.0.0"
description = "A Model Context Protocol server providing tools to convert DOCX to PDF and generate form letters"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Christoph Brill" }]
maintainers = [{ name = "Christoph Brill", email = "opensource@christophbrill.de" }]
keywords = ["libreoffice", "mcp", "llm", "automation"]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "mcp>=1.0.0",
    "fastapi>=0.115.14",
    "uvicorn>=0.34.3",
    "python-multipart>=0.0.20",
    "pydantic>=2.11.7",
    "python-docx>=1.1.2",
    "httpx>=0.24.0",
]

[project.scripts]
libreoffice-mcp = "libreoffice_mcp:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = ["pyright>=1.1.389", "ruff>=0.7.3", "pytest>=8.0.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
